Use canonical <Plug>(argtextobj...) mappings to allow remapping.
Using the argtextobj#Enable() autoload function is counterproductive, as
it triggers the autoload function on VimEnter, negating the startup
speedup of autoload.  It is more natural and flexible to use the
standard <Plug>-mappings instead of a configuration variable to
influence the mappings.

Changes done in a different fork of same upstream repo.
see:  https://github.com/inkarkat/argtextobj.vim/commit/f01ef27